+++ infrastructure/persistence/repositories/observation_repository.py
@@ -5,7 +5,7 @@ from sqlalchemy import and_
 
 from infrastructure.persistence.models import EventObservation
 from infrastructure.persistence.database import db_manager
-from timezone_utils import get_local_now
+from shared.timezone_utils import get_local_now
 
 logger = logging.getLogger(__name__)
