+++ infrastructure/persistence/repositories/odds_repository.py
@@ -4,8 +4,8 @@ from datetime import datetime, timedelta
 
 from infrastructure.persistence.models import OddsSnapshot, EventOdds
 from infrastructure.persistence.database import db_manager
-from timezone_utils import get_local_now
-from odds_utils import validate_odds_data
+from shared.timezone_utils import get_local_now
+from shared.odds_utils import validate_odds_data
 
 logger = logging.getLogger(__name__)
