+++ infrastructure/persistence/repositories/oddsportal_cache_repository.py
@@ -3,7 +3,7 @@ from typing import Optional, Dict
 
 from infrastructure.persistence.models import OddsPortalLeagueCache
 from infrastructure.persistence.database import db_manager
-from timezone_utils import get_local_now
+from shared.timezone_utils import get_local_now
 
 logger = logging.getLogger(__name__)
