+++ oddsportal_config.py
@@ -9,7 +9,7 @@ from typing import Any, Dict, List, Optional
 
 def get_oddsportal_current_date() -> date:
     """Return the local calendar date used for OddsPortal cache decisions."""
-    from timezone_utils import get_local_now
+    from shared.timezone_utils import get_local_now
 
     return get_local_now().date()
