+++ today_sport_extractor.py
@@ -6,8 +6,8 @@ from decimal import Decimal
 import sofascore_api2  # Import to attach methods to SofaScoreAPI class
 from sofascore_api import api_client
 from infrastructure.persistence.repositories import EventRepository, OddsRepository, DailyDiscoveryRepository
-from odds_utils import fractional_to_decimal, validate_odds_data
-from timezone_utils import get_local_now_aware
+from shared.odds_utils import fractional_to_decimal, validate_odds_data
+from shared.timezone_utils import get_local_now_aware
 
 logger = logging.getLogger(__name__)
